Changing the frame id from odom to NED.
The frame id is the reference frame, so when we move, from where are we moving. When the yaw change, it is by reference to what. The submarine works following the NED convention. When visualizing with RViz, the script ENU_to_NED should be running. ROS (by extent, RViz) is in ENU, so we need to convert the base axis reference to NED with the script to visualize correctly the data (X being real X, Z going down, etc.)